.cursor/skills/aap-job-template-automation/requirements.txt
# Python dependencies for AAP Job Template Automation skill

# YAML parsing
pyyaml>=6.0

# HTTP client for AAP API
requests>=2.28.0
